refactor: etxracted calculate price
